+++ External/FEXCore/Source/Interface/Core/OpcodeDispatcher/Flags.cpp
@@ -7,10 +7,10 @@ $end_info$
 
 #include "Interface/Context/Context.h"
 #include "Interface/Core/OpcodeDispatcher.h"
+#include "Interface/Core/X86Tables/X86Tables.h"
 
 #include <FEXCore/Core/X86Enums.h>
 #include <FEXCore/Config/Config.h>
-#include <FEXCore/Debug/X86Tables.h>
 #include <FEXCore/Utils/LogManager.h>
 #include <FEXCore/IR/IR.h>
