+++ External/FEXCore/Source/Interface/Core/X86Tables/BaseTables.cpp
@@ -7,7 +7,6 @@ $end_info$
 #include "Interface/Core/X86Tables/X86Tables.h"
 
 #include <FEXCore/Core/Context.h>
-#include <FEXCore/Debug/X86Tables.h>
 
 #include <iterator>
