+++ External/FEXCore/Source/Interface/Core/X86Tables/H0F38Tables.cpp
@@ -5,7 +5,6 @@ $end_info$
 */
 
 #include "Interface/Core/X86Tables/X86Tables.h"
-#include <FEXCore/Debug/X86Tables.h>
 
 #include <iterator>
 #include <stdint.h>
