+++ External/FEXCore/Source/Interface/Core/X86Tables/SecondaryModRMTables.cpp
@@ -6,8 +6,6 @@ $end_info$
 
 #include "Interface/Core/X86Tables/X86Tables.h"
 
-#include <FEXCore/Debug/X86Tables.h>
-
 #include <iterator>
 
 namespace FEXCore::X86Tables {
