+++ External/FEXCore/Source/Interface/Core/X86Tables/VEXTables.cpp
@@ -6,7 +6,6 @@ $end_info$
 
 #include "Interface/Core/X86Tables/X86Tables.h"
 
-#include <FEXCore/Debug/X86Tables.h>
 #include <iterator>
 
 namespace FEXCore::X86Tables {
